🐛 Fix potential underflow bug in max simulation computation (#144)
## Description

The following computation is currently being used to compute the maximum
number of simulations to run:


https://github.com/cda-tum/qcec/blob/5ae4e6144d9144a128dbb21827960d2169d250ce/include/Configuration.hpp#L69-L70

However, `std::thread::hardware_concurrency` is not guaranteed to return
a meaningful result. From its documentation:

> Number of concurrent threads supported. If the value is not well
defined or not computable, returns ​`0​`.

In the above computation, if `std::thread::hardware_concurrency` happens
to return `0` or reports only a single available thread, the subtraction
would underflow. As a result, the maximum would be set to roughly
`2**64` simulations -- essentially guaranteeing an endless run.

This PR adds an additional sanity check to make sure that the maximum
number of simulations is properly set up.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>