📌 More lenient package requirements (#182)
This small PR relaxes the (lower and upper) caps on QCEC's Python
dependencies in order to increase it's compatibility with other
libraries. This is mainly inspired by
https://iscinumpy.dev/post/bound-version-constraints/.

In the future, it might make sense to introduce a CI check that ensures
the lowest allowed versions still work. Such a job would be helpful for
ensuring that QCEC continues to work with a large variety of external
libraries and hardly depends on them.
Pip `constraint` files could work for that purpose. 

Signed-off-by: burgholzer <burgholzer@me.com>