🔥 Replace weakly maintained `setup-nox` action (#199)
## Description

The `setup-nox` action (https://github.com/daisylb/setup-nox) isn't very
well maintained. This PR drops it in favor of a simple `pipx install
nox` and the corresponding Python setup.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>