👷 Add CI workflows for `s390x` and `ppc64le` architectures (#198)
## Description

This PR adds CI support for building `s390x` and `ppc64le` Python
wheels. This completes the long list of wheels that is offered by
Qiskit. Hence, once this PR is merged and a release is created, QCEC
provides binary wheels for any platform supported by Qiskit.
Since these are emulated platforms and GitHub places a 6h timeout on
jobs, the jobs are split up as much as possible (creating 10 new jobs).
Jobs for these platforms are split into a separate workflow file and
only run on releases.

resolves #153 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>