Fix bug with hadamards in idenitity diagram. (#201)
## Description

Fixes a bug where the ZX-checker simplifies an equivalence checking
miter to a diagram consisting only of wires. If this happens the checker
erroneously didn't check whether the wires are hadamard wires.

Fixes #200 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.