Fix segfault in ZX-checker when idle qubits are stripped (#253)
## Description

Fixes #251.

The problem was that "incomplete" mappings that occur when qubits are
stripped were not handled correctly. Sometimes an index that couldn't
appear as an output qubit in the ZX-diagram miter was used.

However, in the process of fixing this bug, it has become apparent that
the ZX-checker also requires some form of permutation handling like with
the dd checker. This would have to be done as a pre-processing step
because the permutations have to be worked out properly before building
the equivalence checking miter.

The current version doesn't segfault anymore because there is never a
case where qubit indices out of range are assigned. Therefore the
diagrams are always concatenated in some way. Now since they are not
always connected correctly, the ZX-checker just yields NoInformation in
those cases.
## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.