✨ Partial equivalence checking (#375)
## Description

Added a flag `checkPartialEquivalence` in order to decide whether to
check for total equivalence or for partial equivalence. Two circuits are
partially equivalent if, for each possible initial input state, they
have the same probability for each measurement outcome. Therefore the
state of not measured qubits (= garbage qubits), or phases are ignored.

The Construction Checker and the Simulation Checker implement this by
reducing the contribution of garbage qubits in the matrix/vector
representation of the circuit. Additionally, all weights in the DD
representation are set to their magnitude, such that different phases
don't influence the result.

The Alternating Checker implements it by checking that the resulting
matrix representation is equal to a diagonal matrix, modulo garbage
qubits (instead of checking that it is equal to the identity matrix).

For more information, there is documentation in the file
`PartialEquivalence.ipynb`.

An additional feature are the automatically generated benchmarks for
partial equivalence. The function `generatePartiallyEquivalentCircuits`
generates pairs of circuits which are partially equivalent, following
the method described in the paper [Partial Equivalence Checking of
Quantum Circuits](https://arxiv.org/abs/2208.07564) in Section VI. B.


Fixes issue #155 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>