⬆️ update to `mqt-workflows==1.0.1` (#400)
## Description

This PR updates the reusable workflows to their latest versions. This
brings a better Python coverage collection method and an explicit option
to trigger clang-tidy checks on the whole repository.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.