💚 fix RtD build (#584)
## Description

This small PR fixes the RtD build pipeline.
The manylinux containers use a more recent version of gcc (gcc-14) than
the default that comes with Ubuntu 24.04 (gcc-13).
Since `pybind11` hasn't had a new release with the looser compatibility
between libraries compiled with different compilers or compiler
versions, we need to build mqt-core from source for now in order to
build the docs.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>