⬆️ update mqt-core to stable v3 release (#597)
## Description

This PR updates mqt-core to the stable version of the v3 release series.
Notably, this PR pins the build time dependency for mqt-core to the
`3.0.x` minor series as this is also the compatibility ensured by the
mqt-core shared libraries.
This implies that we need to actively update QCEC if we want to build
with mqt-core `3.1+`.
The main reason for this constraint is to minimize unnecessary breaking
changes in the build pipeline.
We do not cap the runtime version of mqt-core to maximize compatibility.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>