🚚 Prepare QCEC for organization move to the `munich-quantum-toolkit` org (#598)
## Description

This PR prepares QCEC for being moved to the new MQT GitHub
organization.
It updates all the respective references as well as license headers.
Additionally, it updates the PR and issue templates.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>